shark(1) ▃▅▃▅▃ Sparkline Generator
A sparkline is a type of line chart that provides a simple
representation of numerical or statistical variation in some
measurement, typically over time.

Shark reads the standard input extracting numeric values and
produces a stream of UTF-8 block characters of increasing
height. Input is read up to NULL and processed as a single
record by default. To process each line as a separate record,
use the `--line` flag.

To process the input in real time, use `--max`=*number* and
`--min`=*number* to set the dataset range in advance.